frontend/requirements.txt
Flask
psycopg2
Werkzeug
google.cloud.storage